requirements.txt
altgraph==0.17.2
async-generator==1.10    
attrs==22.1.0            
certifi==2022.6.15       
cffi==1.15.1             
charset-normalizer==2.1.1
comtypes==1.1.14         
future==0.18.2           
h11==0.13.0
idna==3.3
keyboard==0.13.5
outcome==1.2.0
pefile==2022.5.30
psycopg2==2.9.3
pycparser==2.21
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pywin32==304
pywin32-ctypes==0.2.0
pywinauto==0.6.8
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
wsproto==1.1.0